GPSReader/requirements.txt
paho-mqtt==1.6.1
pynmeagps==1.0.16
pyserial==3.5